requirements.txt
openai
sounddevice
soundfile
numpy
python-dotenv
simpleaudio
openai[realtime]